refactor: update LinkedIn integration and improve code structure
